+++ websockets/py35/client_server.py
@@ -17,17 +17,27 @@ class ClientServerContextManager:
     def tearDown(self):
         self.loop.close()
 
-    def test_basic(self):
+    def test_client(self):
         server = serve(handler, 'localhost', 8642)
         self.server = self.loop.run_until_complete(server)
 
-        async def basic():
+        async def run_client():
             async with connect('ws://localhost:8642/') as client:
                 await client.send("Hello!")
                 reply = await client.recv()
                 self.assertEqual(reply, "Hello!")
 
-        self.loop.run_until_complete(basic())
+        self.loop.run_until_complete(run_client())
 
         self.server.close()
         self.loop.run_until_complete(self.server.wait_closed())
+
+    def test_server(self):
+        async def run_server():
+            async with serve(handler, 'localhost', 8642):
+                client = await connect('ws://localhost:8642/')
+                await client.send("Hello!")
+                reply = await client.recv()
+                self.assertEqual(reply, "Hello!")
+
+        self.loop.run_until_complete(run_server())